% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_data.R
\name{get_port_data}
\alias{get_port_data}
\title{Lists all the ports and their basic information}
\usage{
get_port_data()
}
\value{
A data frame with the following columns: port_number, port_border, port_name, port_crossing_name, port_hours, port_commercial_automation_type, port_commercial_max_lanes, port_passenger_automation_type, port_passenger_max_lanes, port_pedestrian_automation_type, port_pedestrian_max_lanes
}
\description{
Lists all the ports and their basic information
}
