% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_status.R
\name{get_port_status}
\alias{get_port_status}
\title{gets the current status for each port}
\usage{
get_port_status()
}
\value{
a dataframe with the current status for each port
}
\description{
gets the current status for each port
}
