% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waittimes.R
\name{plot_waittimes}
\alias{plot_waittimes}
\title{A histogram plot comparing all wait times to a selected one}
\usage{
plot_waittimes(selected_port = "535503", mod = "passenger", ln_typ = "ready")
}
\arguments{
\item{selected_port}{a port id number}

\item{mod}{the modality of interest}

\item{ln_typ}{the lane type of interest}
}
\value{
a plot
}
\description{
A histogram plot comparing all wait times to a selected one
}
