% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_collector.R
\name{port_collector}
\alias{port_collector}
\title{collects current port status}
\usage{
port_collector(db_name = "portpack.db")
}
\arguments{
\item{db_name}{name of the database (default: "portpack.db")}
}
\description{
collects current port status
}
