% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_list_collector.R
\name{port_list_collector}
\alias{port_list_collector}
\title{collect list of all ports}
\usage{
port_list_collector(db_name = "portpack.db")
}
\arguments{
\item{db_name}{name of the database (default: "portpack.db")}
}
\description{
collect list of all ports
}
