% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_portpack_db.R
\name{portpackdb}
\alias{portpackdb}
\title{initialize port data database}
\usage{
portpackdb(db_name = "portpack.db")
}
\arguments{
\item{db_name}{name of SQLite database}
}
\description{
initialize port data database
}
