% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ports.R
\name{scrape_ports}
\alias{scrape_ports}
\title{Scrape current status of US-Canada border crossings}
\usage{
scrape_ports(xml_url = "https://bwt.cbp.gov/xml/bwt.xml")
}
\arguments{
\item{xml_url}{A character vector with one element, the URL of the XML file (default as of 2/20/2024: "https://bwt.cbp.gov/xml/bwt.xml")}
}
\value{
A data frame with one row per port and one column per attribute from the raw xml dataset
}
\description{
Scrape current status of US-Canada border crossings
}
