% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_collector.R
\name{status_collector}
\alias{status_collector}
\title{collects current port status}
\usage{
status_collector(db_name = "portpack.db")
}
\arguments{
\item{db_name}{name of the database (default: "portpack.db")}
}
\description{
collects current port status
}
