% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into a list of substrings}
\usage{
strsplit1(x, sep)
}
\arguments{
\item{x}{A character vector with one element}

\item{sep}{A character vector indicating what to split on}
}
\value{
A character vector
}
\description{
Split a string into a list of substrings
}
\examples{
x <- c("alpha,bravo,charlie,delta,echo,foxtrot")
strsplit1(x, ",")
}
